+++ src/theme.js
@@ -32,9 +32,9 @@ const theme = {
     black: kradoColors.ash800
   },
   fonts: {
-    body: "'Source Sans Pro', Arial, sans-serif",
-    heading: "'Source Sans Pro', Arial, sans-serif",
-    display: "'Source Serif Pro', Georgia, serif"
+    body: 'Barlow, Arial, sans-serif',
+    heading: 'Barlow, Arial, sans-serif',
+    display: 'Merriweather, Georgia, serif'
   },
   fontWeights: {
     body: 500,